requirements.txt
Flask==3.1.1
flask-cors==6.0.1
openai==1.91.0
gunicorn==23.0.0
python-dotenv==1.0.0 